Refactor - Focused on Backup and Update Functions, Improved Code (#3) (#4)
* [Refactor] Refactoring application: Work in progress

This commit represents an ongoing refactoring process aimed at improving the structure and performance of the application. However, the refactoring is not yet complete.

During this process, various parts of the code have been modified, including enhancements to organization, removal of duplicate code, and optimization of database queries. Additionally, new features have been added to improve the overall functionality of the application.

It is important to note that this commit only represents an initial stage of the refactoring and does not reflect the final state of the application. There is still work to be done to complete the refactoring and ensure that all aspects of the application are properly addressed.

Subsequent commits will continue to address the remaining areas of the refactoring and implement the planned improvements. We ask for your patience and understanding as we work towards delivering a more robust and efficient application.

Please note that this commit may introduce temporary changes or unfinished code. These will be refined and finalized in subsequent commits.

* feat: Add API update function (#2)

This commit adds a new feature that enables updating the API. The update function is responsible for processing and applying available updates to the API, ensuring that the system always uses the latest version and maintains data integrity.

* Refactoring: Prepare structure for update functions

This commit aims to prepare the necessary structure for the implementation of update functions. At this stage, the actual functions have not been created yet, but initial preparations have been made for their development.

Furthermore, Docker configuration has been initiated as part of the development process. Docker integration will enable the creation of an isolated environment to run and test the application, providing greater ease and consistency in the development process.

The next steps will involve implementing the update functions and finalizing the Docker configuration for the application.

These changes aim to improve the structure and lay the groundwork for the addition of update functions, while ensuring a more efficient and consistent development environment through Docker.

* Refactor Dockerfile for improved project setup

This commit refactors the Dockerfile to enhance the project setup process. The following changes have been made:

- Updated system dependencies installation to include additional packages required by the project.
- Added the copying of the `requirements.txt` file to the working directory and installed the project dependencies using `pip`.
- Installed the `pysqlcipher3` library, which is necessary for the project.
- Copied the project source code to the working directory.

These changes optimize the Docker setup, providing a more streamlined and consistent environment for running the project.

Please note that this commit does not introduce any functional changes to the application itself, but focuses solely on improving the Docker setup.

* Refactor: Implement encryption with cipher and improve logging configuration

This commit focuses on refactoring the codebase to improve its structure, organization, and maintainability. One of the key changes is the implementation of data encryption using the cipher algorithm. We have integrated cipher encryption into our SQLite database, enhancing the security of sensitive data.

Additionally, we have made significant improvements to the logging system. The logger has been properly configured and integrated throughout the application to provide more informative and reliable logs.

Please note that there is still work to be done, as we continue to enhance and optimize the system.

* chore: Generate new requirements file

A new requirements file has been generated to reflect the project dependencies. This file includes all the necessary packages and versions required for the project.

Committing the updated requirements file ensures that the project can be easily replicated and that all dependencies are properly managed.

* [fix] Update requirements.txt and Dockerfile
[chore] Ongoing troubleshooting for container and image execution

* feat(authentication, backup): add initial route authentication and backup function; refactor existing codebase

The codebase has been refactored for improved readability and maintainability. An initial implementation of a backup function and route authentication has been added, but is not yet complete.

* refactor: Fix registration and authentication in blueprint routes

This commit refactors the code to address the registration and authentication issue in the blueprint routes. The registration of blueprints has been adjusted to ensure correct routing and prevent conflicts. The authentication process, including the JWT decorator usage, has been revised to properly handle authentication in the protected routes. This resolves the issue and improves the overall functionality of the application.

* feat: Update backup functions and .gitignore

Backup functions have been updated and are now working correctly. The .gitignore file has also been updated to ignore the backup folder. For security reasons, authentication has been temporarily removed and will be reinstated at a later stage in the project.

* feat: implemented update functions

Implemented update functions